Create rosbridge-server subprocess on startup
This will not work if you do not have rosbridge_server installed.
Also, as of now the rosbridge server has a longer startup-time
than the application, therefore the error message will show for
a brief second. This should be fixed, for example by creating the
application window when the rosbridge-server is ready.